Add MIT License to project
There was no explicit license to this particular project so I looked at
the other projects which are owned by Pusher, Ltd and found that the MIT
license was mainly used.

This commit brings over the MIT License to this project to make it clear
that it is an open sourced project which can be used freely instead of
leaving that up for interpretation to users.